+++ src/rule-definitions.ts
@@ -3,11 +3,11 @@ import banUnboundedStringType from '#src/rules/ban-unbounded-string-type.js';
 import enumNamePascalCase from '#src/rules/enum-name-pascal-case.js';
 import enumValueSnakeCase from '#src/rules/enum-value-snake-case.js';
 import fieldNameCamelCase from '#src/rules/field-name-camel-case.js';
+import fieldNameGrammaticalNumber from '#src/rules/field-name-grammatical-number.js';
 import fieldNameMappingSnakeCase from '#src/rules/field-name-mapping-snake-case.js';
 import fieldOrder from '#src/rules/field-order.js';
 import forbidField from '#src/rules/forbid-field.js';
 import forbidRequiredIgnoredField from '#src/rules/forbid-required-ignored-field.js';
-import listFieldNameGrammaticalNumber from '#src/rules/list-field-name-grammatical-number.js';
 import modelNameGrammaticalNumber from '#src/rules/model-name-grammatical-number.js';
 import modelNameMappingSnakeCase from '#src/rules/model-name-mapping-snake-case.js';
 import modelNamePascalCase from '#src/rules/model-name-pascal-case.js';
@@ -22,11 +22,11 @@ export default [
   enumNamePascalCase,
   enumValueSnakeCase,
   fieldNameCamelCase,
+  fieldNameGrammaticalNumber,
   fieldNameMappingSnakeCase,
   fieldOrder,
   forbidField,
   forbidRequiredIgnoredField,
-  listFieldNameGrammaticalNumber,
   modelNameGrammaticalNumber,
   modelNameMappingSnakeCase,
   modelNamePascalCase,